refactor(mypage): 찜 목록 전용 타입 위치 정리
