--- src/features/mypage/types.ts
@@ -1,7 +1,0 @@
-export interface FavoriteGamePreview {
-  gameId: number;
-  title: string;
-  summary: string;
-  thumbnailUrl: string | null;
-  genres: string[];
-}
