+++ src/pages/mypage/types.ts
@@ -1,3 +1,11 @@
+export interface FavoriteGamePreview {
+  gameId: number;
+  title: string;
+  summary: string;
+  thumbnailUrl: string | null;
+  genres: string[];
+}
+
 export type MyPageToast = {
   tone: 'success' | 'error';
   message: string;
